refactor: rename useToDetailPage to useToMarketDetailPage for clarity
- Renamed hook from useToDetailPage to useToMarketDetailPage to better reflect its purpose
- Updated imports in MarketTokenListBase and UniversalSearchV2MarketTokenItem
- Maintains same functionality with improved naming convention
